Moved unit tests to dedicated directory
Also created constants for the locations of the fixtures and mocks for testing so we do not have to use relative paths in tests
