+++ phpunit.xml.dist
@@ -6,7 +6,7 @@
         xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/7.4/phpunit.xsd"
         backupGlobals="false"
         colors="true"
-        bootstrap="vendor/autoload.php">
+        bootstrap="tests/bootstrap.php">
     <php>
         <ini name="error_reporting" value="-1" />
         <ini name="date.timezone" value="UTC" />
