+++ tests/Mocks/responses/stable-body.txt
