+++ tests/bootstrap.php
@@ -1,0 +1,6 @@
+<?php declare(strict_types=1);
+
+require_once __DIR__ . '/../vendor/autoload.php';
+
+define('TEST_FIXTURES_DIRECTORY', __DIR__ . '/Fixtures');
+define('TEST_MOCKS_DIRECTORY', __DIR__ . '/Mocks');
